cross-compiled builds
Enable cross-compiled builds for multiple platforms. Build for all known
platforms by default, but allow the user to specify a list of platforms
to be used instead. The process is:

* Clone the repo on BUILDBRANCH, determine the most recent release tag
  on that branch, and checkout that tag.
* For each specified platform:
  - Check if that build has already happened. If so, move on
  - Build
  - Upload
  - Cut release

If any of the individual platform builds fail, an error is logged, but
the build process continues for other platforms.

In order to guarantee reproducible builds on any platform, set
`CGO_ENABLED=0` to ensure that we always build statically-linked
binaries.